agihouse-hackathon/pyproject.toml
[project]
name = "agihouse-hackathon-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "google-genai>=0.5.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
]
